clean main.go
